+++ packages/react-components/react-headless-components-preview/stories/src/MessageBar/index.stories.tsx
@@ -3,10 +3,9 @@ import {
   MessageBarActions,
   MessageBarBody,
   MessageBarTitle,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/message-bar';
 
 import descriptionMd from './MessageBarDescription.md';
-
 export { Default } from './MessageBarDefault.stories';
 export { Intent } from './MessageBarIntent.stories';