+++ packages/react-components/react-headless-components-preview/stories/src/Toolbar/index.stories.tsx
@@ -5,10 +5,9 @@ import {
   ToolbarGroup,
   ToolbarRadioGroup,
   ToolbarToggleButton,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/toolbar';
 
 import descriptionMd from './ToolbarDescription.md';
-
 export { Default } from './ToolbarDefault.stories';
 export { Vertical } from './ToolbarVertical.stories';
 export { Toggle } from './ToolbarToggleButton.stories';