+++ packages/react-components/react-motion-components-preview/stories/src/Collapse/Collapse.stories.tsx
@@ -1,7 +1,13 @@
 import * as React from 'react';
 import type { JSXElement } from '@fluentui/react-components';
-import type { PresenceComponentProps } from '@fluentui/react-components';
-import { Collapse } from '@fluentui/react-motion-components-preview';
+import { makeStyles, tokens, type PresenceComponentProps } from '@fluentui/react-components';
+import { Collapse, type CollapseParams } from '@fluentui/react-motion-components-preview';
+
+const useClasses = makeStyles({
+  wrapper: {
+    padding: tokens.spacingVerticalXL,
+  },
+});
 
 const LoremIpsum = React.forwardRef<HTMLDivElement, React.HTMLProps<HTMLDivElement>>((props, ref) => (
   <div ref={ref} {...props}>
@@ -11,10 +17,13 @@ const LoremIpsum = React.forwardRef<HTMLDivElement, React.HTMLProps<HTMLDivEleme
   </div>
 ));
 
-export const DefaultCollapse = (props: PresenceComponentProps): JSXElement => {
+export const DefaultCollapse = (props: PresenceComponentProps & CollapseParams): JSXElement => {
+  const classes = useClasses();
   return (
-    <Collapse {...props}>
-      <LoremIpsum />
-    </Collapse>
+    <div className={classes.wrapper}>
+      <Collapse {...props}>
+        <LoremIpsum />
+      </Collapse>
+    </div>
   );
 };