+++ packages/web-components/src/menu-list/menu-list.definition.ts
@@ -1,4 +1,4 @@
-import { FluentDesignSystem } from '../fluent-design-system.js';
+import { tagName } from './menu-list.options.js';
 import { MenuList } from './menu-list.js';
 import { styles } from './menu-list.styles.js';
 import { template } from './menu-list.template.js';
@@ -9,7 +9,7 @@ import { template } from './menu-list.template.js';
  * HTML Element: <fluent-menu-list>
  */
 export const definition = MenuList.compose({
-  name: `${FluentDesignSystem.prefix}-menu-list`,
+  name: tagName,
   template,
   styles,
 });