+++ packages/web-components/src/tab/tab.definition.ts
@@ -1,10 +1,10 @@
-import { FluentDesignSystem } from '../fluent-design-system.js';
+import { tagName } from './tab.options.js';
 import { Tab } from './tab.js';
 import { template } from './tab.template.js';
 import { styles } from './tab.styles.js';
 
 export const definition = Tab.compose({
-  name: `${FluentDesignSystem.prefix}-tab`,
+  name: tagName,
   template,
   styles,
 });