+++ packages/react-components/react-icons-compat/library/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-icons-compat
 
-This log was last generated on Wed, 25 Feb 2026 13:32:22 GMT and should not be manually modified.
+This log was last generated on Thu, 23 Apr 2026 14:21:16 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.2.19](https://github.com/microsoft/fluentui/tree/@fluentui/react-icons-compat_v0.2.19)
+
+Thu, 23 Apr 2026 14:21:16 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-icons-compat_v0.2.18..@fluentui/react-icons-compat_v0.2.19)
+
+### Patches
+
+- Bump @fluentui/react-jsx-runtime to v9.4.2 ([PR #36035](https://github.com/microsoft/fluentui/pull/36035) by beachball)
+- Bump @fluentui/react-utilities to v9.26.3 ([PR #36035](https://github.com/microsoft/fluentui/pull/36035) by beachball)
 
 ## [0.2.18](https://github.com/microsoft/fluentui/tree/@fluentui/react-icons-compat_v0.2.18)